chore(Method) : 각 인터페이스에 주석으로 설명
각 메서드가 무슨 역할을 하게 될 것인지에 대한 설명을 주석으로 작성함
